+++ src/whatsappConstants.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+POLL_CACHE_VERSION = 1
+RECENT_POLLS_TO_RECHECK = 2
+IGNORE_POLL_CACHE = True  # temporary while stabilising poll discovery
